+++ api/src/prisma/schema/migrations/20250616154411_defined_role_model/migration.sql
@@ -1,0 +1,42 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `role` on the `User` table. All the data in the column will be lost.
+
+*/
+-- CreateEnum
+CREATE TYPE "RoleNames" AS ENUM ('VIEWER', 'AUTHOR', 'ADMIN');
+
+-- AlterTable
+ALTER TABLE "User" DROP COLUMN "role";
+
+-- DropEnum
+DROP TYPE "Role";
+
+-- CreateTable
+CREATE TABLE "Role" (
+    "id" TEXT NOT NULL,
+    "name" "RoleNames" NOT NULL DEFAULT 'VIEWER',
+
+    CONSTRAINT "Role_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateTable
+CREATE TABLE "_Roles" (
+    "A" TEXT NOT NULL,
+    "B" TEXT NOT NULL,
+
+    CONSTRAINT "_Roles_AB_pkey" PRIMARY KEY ("A","B")
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Role_id_key" ON "Role"("id");
+
+-- CreateIndex
+CREATE INDEX "_Roles_B_index" ON "_Roles"("B");
+
+-- AddForeignKey
+ALTER TABLE "_Roles" ADD CONSTRAINT "_Roles_A_fkey" FOREIGN KEY ("A") REFERENCES "Role"("id") ON DELETE CASCADE ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "_Roles" ADD CONSTRAINT "_Roles_B_fkey" FOREIGN KEY ("B") REFERENCES "User"("id") ON DELETE CASCADE ON UPDATE CASCADE;