+++ api/src/prisma/schema/migrations/20250616173608_defined_roles_model/migration.sql
@@ -1,0 +1,28 @@
+/*
+  Warnings:
+
+  - You are about to drop the `_Roles` table. If the table is not empty, all the data it contains will be lost.
+
+*/
+-- DropForeignKey
+ALTER TABLE "_Roles" DROP CONSTRAINT "_Roles_A_fkey";
+
+-- DropForeignKey
+ALTER TABLE "_Roles" DROP CONSTRAINT "_Roles_B_fkey";
+
+-- DropTable
+DROP TABLE "_Roles";
+
+-- CreateTable
+CREATE TABLE "Roles" (
+    "userId" TEXT NOT NULL,
+    "roleId" TEXT NOT NULL,
+
+    CONSTRAINT "Roles_pkey" PRIMARY KEY ("userId","roleId")
+);
+
+-- AddForeignKey
+ALTER TABLE "Roles" ADD CONSTRAINT "Roles_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "Roles" ADD CONSTRAINT "Roles_roleId_fkey" FOREIGN KEY ("roleId") REFERENCES "Role"("id") ON DELETE RESTRICT ON UPDATE CASCADE;