+++ api/src/prisma/schema/migrations/20250616202248_added_field_assigned_at/migration.sql
@@ -1,0 +1,34 @@
+/*
+  Warnings:
+
+  - The `name` column on the `Role` table would be dropped and recreated. This will lead to data loss if there is data in the column.
+  - A unique constraint covering the columns `[name]` on the table `Role` will be added. If there are existing duplicate values, this will fail.
+
+*/
+-- CreateEnum
+CREATE TYPE "RoleName" AS ENUM ('VIEWER', 'AUTHOR', 'ADMIN');
+
+-- DropForeignKey
+ALTER TABLE "Roles" DROP CONSTRAINT "Roles_roleId_fkey";
+
+-- DropForeignKey
+ALTER TABLE "Roles" DROP CONSTRAINT "Roles_userId_fkey";
+
+-- AlterTable
+ALTER TABLE "Role" DROP COLUMN "name",
+ADD COLUMN     "name" "RoleName" NOT NULL DEFAULT 'VIEWER';
+
+-- AlterTable
+ALTER TABLE "Roles" ADD COLUMN     "assignedAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP;
+
+-- DropEnum
+DROP TYPE "RoleNames";
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Role_name_key" ON "Role"("name");
+
+-- AddForeignKey
+ALTER TABLE "Roles" ADD CONSTRAINT "Roles_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE CASCADE ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "Roles" ADD CONSTRAINT "Roles_roleId_fkey" FOREIGN KEY ("roleId") REFERENCES "Role"("id") ON DELETE CASCADE ON UPDATE CASCADE;